Move expedited.{cpp,h} to thinblock.{cpp,h}
