+++ src/.formatted-files
@@ -9,6 +9,4 @@ parallel.h
 unlimited.cpp
 unlimited.h
 globals.cpp
-expedited.cpp
-expedited.h
 test/thinblock_util_tests.cpp
