+++ src/thinblock.h
@@ -226,4 +226,38 @@ extern std::set<uint256> setUnVerifiedOrphanTxHash;
 extern CCriticalSection cs_xval;
 // Xpress Validation: end
 
+enum
+{
+    EXPEDITED_STOP = 1,
+    EXPEDITED_BLOCKS = 2,
+    EXPEDITED_TXNS = 4,
+};
+
+enum
+{
+    EXPEDITED_MSG_HDR = 1,
+    EXPEDITED_MSG_XTHIN = 2,
+};
+
+
+extern CCriticalSection cs_xpedited; // protects xpeditedBlk, xpeditedBlkUp and xpeditedTxn
+extern std::vector<CNode *> xpeditedBlk; // Who requested expedited blocks from us
+extern std::vector<CNode *> xpeditedBlkUp; // Who we requested expedited blocks from
+extern std::vector<CNode *> xpeditedTxn;
+
+// Checks to see if the node is configured in bitcoin.conf to
+bool CheckAndRequestExpeditedBlocks(CNode *pfrom);
+
+// be an expedited block source and if so, request them.
+void SendExpeditedBlock(CXThinBlock &thinBlock, unsigned char hops, const CNode *skip = NULL);
+void SendExpeditedBlock(const CBlock &block, const CNode *skip = NULL);
+void HandleExpeditedRequest(CDataStream &vRecv, CNode *pfrom);
+bool IsRecentlyExpeditedAndStore(const uint256 &hash);
+
+// process incoming unsolicited block
+bool HandleExpeditedBlock(CDataStream &vRecv, CNode *pfrom);
+
+// is this node an expedited node
+bool IsExpeditedNode(const CNode *pfrom);
+
 #endif // BITCOIN_THINBLOCK_H
